# openapi.yaml
openapi: 3.1.0
info:
  title: Hooter Backend API
  version: '1.0.0'
  description: Internal API for Hooter backend
servers:
  - url: https://hooter_staging.h0oter.com
    description: Staging server

  - url: http://localhost:8888
    description: Local development server
...
paths:
...
  /catalog/products/exists:
    get:
      tags:
        - Catalog

      summary: Check catalog availability

      description: |
        Returns whether the current brand has at least one catalog product.
        This endpoint is intended for determining whether to show the dashboard
        or prompt the user to create their first catalog.

      security:
        - cookieAuth: []

      responses:
        "200":
          description: Catalog availability status.
          content:
            application/json:
              schema:
                type: object
                required:
                  - catalog
                properties:
                  catalog:
                    type: string
                    enum:
                      - available
                      - unavailable

              examples:
                catalog_available:
                  summary: Brand has catalog
                  value:
                    catalog: available

                catalog_unavailable:
                  summary: Brand has no catalog
                  value:
                    catalog: unavailable
        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
        "default":
          $ref: "#/components/responses/default"
  /catalog/products/single:
    post:
      tags:
        - Catalog
      summary: Upload a single catalog
      description: Uploads a single catalog to the Hooter backend server.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/categoryID"
        - $ref: '#/components/parameters/csrf'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SingleCatalogPayload"

      responses:
        "200":
          description: Successfully added a product into the catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: successful
                  message:
                    type: string
                    example:  added the single catalog
                  usku_id:
                    type: string
                    example:  8a985fd1c8704998b82eb8458577d42b
        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/default"
              examples:
                invalid_listing:
                  summary: Invalid lsiting attribute
                  value:
                    status: bad request
                    message: invalid listing attributes
                invalid_category:
                  summary: Invalid category attribute
                  value:
                    status: bad request
                    message: invalid category attributes
        "default":
          $ref: "#/components/responses/default"
  /catalog/products/bulk:
    post:
      summary:  Upload bulk product
      description:  Upload bulk catalog product sheet as per the category and run a get an SSE link to 
        fetch whether products finished uploading.
      tags:
        - Catalog
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/categoryID"
        - $ref: '#/components/parameters/csrf'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - sheet
              properties:
                sheet:
                  type: string
                  format: binary
                  description: Excel (.xlsx) file
      responses:
        "200":
          description: Sheet received to upload products
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: successful
                  message:
                    type: string
                    example: XLSX sheet receieved to upload the products
                  sse_url:
                    type: string
                    example: /catalog/products/events/2xxxxdfffffff
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: successful
                  message:
                    type: string
                    example: XLSX sheet receieved to upload the products
                  sse_url:
                    type: string
                    example: /catalog/products/events/2xxxxdfffffff
  /catalog/products/error_sheet/{job_id}:
    get:
      summary:  Download the xlsx sheet error sheet
      description:  Download the xlsx error sheet if there is any error sheet created for the job id for the 
        product upload event created from the uploaded excel sheet for bulk product upload
      security:
        - cookieAuth: []
      tags:
        - Catalog
      parameters:
        - in: path
          required: true
          name: job_id
          description:  Unique job ID for the bulk product upload
          schema:
            type: string
      responses:
        "200":
          description: HTTP file download response for error sheet
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

        "202":
          description: Bulk product upload has not completed yet
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: pending
                  message:
                    type: string
                    example: hob has not finsihed yet
        "404":
          description:  resource not found

        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
  /catalog/products/counts:
    get:
      summary: Get counts of uploaded catalog
      description: Get the counts of uploaded catalog for total catalog, pending  and completed status catalog
      tags:
        - Catalog
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties: 
                  count:
                    type: object
                    properties: 
                      completed:
                        type: integer
                      pending:
                        type: integer
                      total:
                        type: integer
              example:
                count:
                  completed: 2
                  pending: 22
                  total: 24
        "500":
          description: Could not fetch the data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/default"
              example:
                status: failed
                message: Could not fetch the catalog data
        "default":
          $ref: "#/components/responses/default"
  /catalog/products:
    get:
      summary:  Fetches the uploaded catalog
      description:  Get product lists uploaded in the catalog and the count of uploaded products according
        to the status
      security:
        - cookieAuth: []
      tags:
        - Catalog
      responses:
        "200":
          description:  Successful request
          content:
            application/json:
              schema:
                type: object
                properties:
                  catalog_list:
                    type: array
                    items:
                      type: object
                      properties:
                        usku_id:
                          type: string
                        sku_id:
                          type: string
                        image_url:
                          type: string
                        product_title:
                          type: string
                        product_price:
                          type: integer
                        compared_price:
                          type: integer
                        purchasing_cost:
                          type: integer
                        status:
                          type: string
                          enum: ['completed', 'pending']
                        product_type:
                          type: string
                        type_id:
                          type: string
              example:
                catalog_list:
                  - usku_id: USKU-1001
                    sku_id: SKU-RED-001
                    image_url: /images/red-shirt.jpg
                    product_title: Premium Red T-Shirt
                    product_price: 999
                    compared_price: 1299
                    purchasing_cost: 550
                    status: completed
                    product_type: apparel
                    type_id: TSHIRT001      

                  - usku_id: USKU-1002
                    sku_id: SKU-BLK-042
                    image_url: /images/black-hoodie.jpg
                    product_title: Classic Black Hoodie
                    product_price: 1899
                    compared_price: 2499
                    purchasing_cost: 1100
                    status: completed
                    product_type: apparel
                    type_id: HOODIE042
        "500":
          description:
            Error fetching the catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
                    example:  could not fetch the catalog data
        "default":
          $ref: "#/components/responses/default"
  /catalog/products/{usku_id}:
    get:
      summary:  Fetch product data
      description: Fetch product listing and category
      security:
        - cookieAuth: []
      parameters:
        - name: usku_id
          in: path
          required: true
          description: product unique usku_id
          schema:
            type: string
            example: usku_xxxx_xxx_xxxxxxxx
      tags:
        - Catalog
      responses:
        "200":
          description: Successful request
          content:
            application/json:
              schema:
                type: object
                properties:
                  sku_id:
                    type: string
                  product_title:
                    type: string
                  product_price:
                    type: string
                  product_desc:
                    type: string
              example:
                sku_id: "#sku01"
                product_title:  Blue shirt
                product_description:  Blue shirt with black strips on the right side size(XL)
                product_price:  500.00
        "400":
          description:  Request failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
              example:
                status: failed
                message:  request failed
        "401":
          $ref: "#/components/responses/login_required"
        "403":
          $ref: "#/components/responses/brand_required"
    delete:
      tags:
        - Catalog
      summary: Delete a product
      description: Deletes a product using its unique USKU ID.

      security:
        - cookieAuth: []

      parameters:
        - name: usku_id
          in: path
          required: true
          description: Unique SKU ID of the product to delete.
          schema:
            type: string
        - $ref: '#/components/parameters/csrf'

      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
              example:
                status: success
                message: Product deleted successfully

        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/default"
              example:
                status: invalid request
                message: usku-id not provided

        "401":
          $ref: "#/components/responses/login_required"

        "403":
          $ref: "#/components/responses/brand_required"

        "default":
          $ref: '#/components/responses/default'
    put:
      summary:  Update catalog product
      description:  Update catalog product listing and category attributes data
      security:
        - cookieAuth: []
      tags:
        - Catalog
      parameters:
        - $ref: '#/components/parameters/csrf'
        - name: usku_id
          in: path
          description:  Universal SKU ID of the product
          required: true
          schema:
            type: string
        - name: type-id
          in: query
          description:  Category ID for the product
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description:  Product attribute field and value payload
        content: 
          application/json:
            schema:
              type: object
              properties:
                listing_attributes:
                  type: object
                  properties:
                    sku_id:
                      type: string
                    price:
                      type: string
                    net_weight_kg:
                      type: string
                category_attributes:
                  type: object
                  properties:
                    stitch:
                      type: string
                      enum: ['stitched', 'unstitched']
                    blouse_piece_included:
                      type: string
                      enum: ['included', 'not included']
            example:
              listing_attributes:
                sku_id: '#012'
                price: 1500
                net_weight_kg: 0.75
              category_attributes:
                stitch: 'stitched'
                blouse_piece_included: included
                
      responses:
        "200":
          description:  Successfully updated the product values from the database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/default"
              example:
                status: successful
                message: successfully updated the catalog product
        "default":
          $ref: "#/components/responses/default"
...
  /catalog/products/uploaded-categories:
    get:
      summary: Get uploaded product categories
      description: Get the unique categories of the products which the brand has uploaded
      tags: 
        - Catalog
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status:
                    type: string
                  categories:
                    type: array
                    items:
                      type: object
                      properties: 
                        id:
                          type: string
                        category:
                          type: string
              example:
                status: successful
                categories:
                  - id: gid://shopify/TaxonomyCategory/aa-1-23-2-1
                    category: Saris
        "500":
          description: Failed to fetch the categories
          content:
            application/json:
              schema:
                type: object
                properties: 
                  status:
                    type: string
                  message:
                    type: string
              example:
                status: failed
                message:  Could not fetch the uploaded product categories
        "default":
          $ref: "#/components/responses/default"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    SingleCatalogPayload:
      type: object
      required:
        - listing_attributes
        - category_attributes
      properties:
        listing_attributes:
          $ref: "#/components/schemas/ListingAttributes"

        category_attributes:
          type: object
          description: >
            Category-specific attributes. Mandatory fields are obtained from the
            category attributes API. Additional custom attributes are allowed.
          additionalProperties: true
          properties:
            blouse_fabric:
              type: string
            blouse_length:
              type: string
            blouse_piece_included:
              type: string
            border_type:
              type: string
            brand_color:
              type: string
            care:
              type: string
            color:
              type: string
            fabric:
              type: string
            key_points:
              type: array
              items:
                type: string
            occasion:
              type: string
            other_info:
              type: string
            pallu_type:
              type: string
            pattern:
              type: string
            product_description:
              type: string
            saree_length_m:
              type: string
            stitch:
              type: string
            style:
              type: string
            uid:
              type: string
            work:
              type: string
    ListingAttributes:
      type: object
      required:
        - sku_id
        - title
        - price
      properties:
        sku_id:
          type: string
        title:
          type: string
        price:
          type: number
        compared_price:
          type: number
        purchasing_cost:
          type: number
        vendor:
          type: string
        ean:
          type: string
        hsn:
          type: string
        net_weight:
          type: number
        dead_weight:
          type: number
        volumetric_weight:
          type: number
        brand_name:
          type: string
    default:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
...
    loginRequried:
      type: object
      properties:
        status:
          type: string
          example: restricted
        message:
          type: string
          example: "user is not logged in"
    brandRequired:
      type: object
      properties:
        status:
          type: string
          example: restricted
        message:
          type: string
          example: no brand found for this user
...
  responses:
    login_required: # 401
        description: Login session required
        content: 
          application/json:
            schema:
              $ref : "#/components/schemas/loginRequried"
    brand_required: # 403
        description: Brand session required
        content: 
          application/json:
            schema:
              $ref : "#/components/schemas/brandRequired"
...
    "default":
      description: Default response
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - message
            properties:
              status:
                type: string
              message:
                type: string
  parameters:
...
    categoryID:
      name: type-id
      in: query
      required: true
      schema:
        type: string
      description: Shopify taxonomy category ID
...
    csrf:
        name: X-CSRF-Token
        in: header
        required: true
        description: CSRF acesss token acquired using /security/csrf API
        schema:
          type: string